backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.3.0
anthropic==0.34.0
apscheduler==3.10.4
sqlalchemy==2.0.36
aiosqlite==0.20.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.0
